newproject/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-restframework==0.0.1
djangorestframework==3.14.0
pillow==10.2.0
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
